chore: refactor structure dcr
