+++ src/apps/DisciplineReleaseControl/Components/3D/3dView.tsx
@@ -7,7 +7,7 @@ import { useQuery } from 'react-query';
 import { TagMap, TagOverlay } from '../../../../packages/ModelViewer/components/tagOverlay';
 import { EleNetwork } from '../../Types/eleNetwork';
 import { Pipetest } from '../../Types/pipetest';
-import { getEleNetworks } from '../Electro/getEleNetworks';
+import { getEleNetworks } from '../../utils/api/getEleNetworks';
 import { MessageWrapper, ThreeDModel } from './3dViewStyles';
 import { getIconName, getStatusColor, getTagOverlay as getElectroTagOverlay } from './Helpers';
 import { ElectroIcon } from './icons/ElectroIcon';
