+++ src/apps/DisciplineReleaseControl/__tests__/pipetestStatusHelpers.test.ts
@@ -13,7 +13,7 @@ import {
     isCheckListStepsInRightOrder,
     isCheckListTestOk,
     isPipetestProcessDoneInRightOrder,
-} from '../Functions/statusHelpers';
+} from '../utils/helpers/statusHelpers';
 import { testData1 } from '../Mock/mockData';
 import {
     CheckListStatus,
