+++ src/apps/DisciplineReleaseControl/utils/config/Datasource/responseParser.ts
@@ -1,0 +1,7 @@
+import { chewPipetestDataFromApi } from '../../helpers/statusHelpers';
+
+export const responseParser = async (response: Response) => {
+    let json = JSON.parse(await response.text());
+    json = chewPipetestDataFromApi(json);
+    return json;
+};
