+++ src/apps/DisciplineReleaseControl/utils/config/index.ts
@@ -2,3 +2,5 @@ export * from './Filter/filterConfig';
 export * from './Garden/gardenConfig';
 export * from './Presets/presetConfig';
 export * from './Table/tableConfig';
+export * from './Datasource';
+export * from './StatusBar/statusBarConfig';
